models/medcat_service/requirements.txt
flask>=2.0.0
medcat>=1.2.0
pyyaml>=5.0
